Introduce peer management services
IPeerSelector - This service returns a set of peers that the current node should attempt to connect to. It will try to keep the mesh complete at all times.

IPeerStatusTracker - This service keeps track of, and provides access to the most recent status updates of all peers, and which peers are connected to this node.