Optimize message bus
MessageBus, while extremely fast, was allocating a lot of memory.
To avoid this, we changed a lot of things to ValueTask, as that was the major source of memory usage.

This, however, made performance... not great. So everything was changed so that subscriptions could choose if they wanted targeted methods, broadcasts, or both. This improved things quite a bit.